Moved implementation into .cpp files, add headers to prevent C2039 error
