+++ include/faker-cxx/Helper.h
@@ -7,9 +7,9 @@
 #include <string_view>
 #include <vector>
 
-#include "Number.h"
+#include "../../src/common/LuhnCheck.h"
 #include "../src/common/StringHelper.h"
-#include "LuhnCheck.h"
+#include "Number.h"
 
 namespace faker
 {
@@ -76,21 +76,7 @@ public:
      * Helper::replaceSymbolWithNumber("Your pin is: !####") // "29841"
      * @endcode
      */
-    static std::string replaceSymbolWithNumber(std::string str, const char & symbol = '#')
-    {
-        for (char& ch : str)
-        {
-            if(ch == symbol)
-            {
-                ch = static_cast<char>(Number::integer(0,9) + '0');
-            }
-            else if(ch == '!')
-            {
-                ch = static_cast<char>(Number::integer(2,9) + '0');
-            }
-        }
-        return str;
-    }
+    static std::string replaceSymbolWithNumber(std::string str, const char & symbol = '#');
 
     /**
      * @brief Returns credit card schema with replaced symbols and patterns in a credit card  including Luhn checksum
@@ -108,22 +94,7 @@ public:
      * Helper::replaceCreditCardSymbols("1234-[4-9]-##!!-L") // "1234-9-5298-2"
      * @endcode
      */
-    static std::string replaceCreditCardSymbols(const std::string& inputString = "6453-####-####-####-###L", char symbol = '#') {
-        // Replace regex-like expressions in the given string with matching values.
-        std::string modifiedString = regexpStyleStringParse(inputString);
-
-        // Replace ### with random numbers (where ### is the symbol)
-        modifiedString = replaceSymbolWithNumber(modifiedString, symbol);
-
-        // Calculate the luhnCheckValue and replace 'L' with the checkNum
-        int checkNum = LuhnCheck::luhnCheckValue(modifiedString);
-        size_t pos = modifiedString.find('L');
-        if (pos != std::string::npos) {
-            modifiedString.replace(pos, 1, std::to_string(checkNum));
-        }
-
-        return modifiedString;
-    }
+    static std::string replaceCreditCardSymbols(const std::string& inputString = "6453-####-####-####-###L", char symbol = '#');
 
     /**
      * @brief Returns the replaced regex-like expression in the string with matching values.
@@ -145,86 +116,7 @@ public:
      * Helper::regexpStyleStringParse("#{3}test[1-5]") // "###test3"
      * @endcode
      */
-    static std::string regexpStyleStringParse(const std::string& input) {
-        std::string string = input;
-        // Deal with range repeat `{min,max}`
-        const std::regex RANGE_REP_REG(R"((.)\{(\d+),(\d+)\})");
-        const std::regex REP_REG(R"((.)\{(\d+)\})");
-        const std::regex RANGE_REG(R"(\[(\d+)-(\d+)\])");
-        std::smatch token;
-
-        while (std::regex_search(string, token, RANGE_REP_REG)) {
-            int min = std::stoi(token[2]);
-            int max = std::stoi(token[3]);
-            // switch min and max
-            if (min > max) {
-                std::swap(min, max);
-            }
-
-            int repetitions = Number::integer(min, max);
-            string =
-                string.substr(0, static_cast<unsigned long>(token.position())) +
-                StringHelper::repeat(token[1], repetitions) +
-                string.substr(static_cast<unsigned long>(token.position() + token.length()));
-        }
-
-        // Deal with repeat `{num}`
-        while (std::regex_search(string, token, REP_REG)) {
-            int repetitions = std::stoi(token[2]);
-            string =
-                string.substr(0, static_cast<unsigned long>(token.position())) +
-                StringHelper::repeat(token[1], repetitions) +
-                string.substr(static_cast<unsigned long>(token.position() + token.length()));
-        }
-
-        // Deal with range `[min-max]` (only works with numbers for now)
-        // TODO: implement for letters e.g. [0-9a-zA-Z] etc.
-        while (std::regex_search(string, token, RANGE_REG)) {
-            int min = std::stoi(token[1]); // This time we are not capturing the char before `[]`
-            int max = std::stoi(token[2]);
-            // switch min and max
-            if (min > max) {
-                std::swap(min, max);
-            }
-
-            string =
-                string.substr(0, static_cast<unsigned long>(token.position())) +
-                std::to_string(Number::integer(min, max)) +
-                string.substr(static_cast<unsigned long>(token.position() + token.length()));
-        }
-
-        return string;
-    }
-
-//    /**
-//     * @brief Returns the luhn check value for the given string.
-//     *
-//     * @param inputString The string to calculate the check value for.
-//     *
-//     * @return the luhn check value for the given string.
-//     *
-//     */
-//    static int luhnCheckValue(const std::string& inputString) {
-//        std::string str = inputString;
-//        str = std::regex_replace(str, std::regex("[\\s-]"), "");
-//        int sum = 0;
-//        bool alternate = false;
-//
-//        for (std::string::size_type i = str.length() - 1; i != std::string::npos; i--) {
-//            int n = str[i] - '0';
-//            if (alternate) {
-//                n *= 2;
-//                if (n > 9) {
-//                    n = (n % 10) + 1;
-//                }
-//            }
-//
-//            sum += n;
-//            alternate = !alternate;
-//        }
-//
-//        return sum % 10;
-//    }
+    static std::string regexpStyleStringParse(const std::string& input);
 
 private:
     static std::random_device randomDevice;
